make CostCalculator trait so it can be better tested
