+++ media-api/app/lib/usagerights/CostCalculator.scala
@@ -1,11 +1,11 @@
 package lib.usagerights
 
-import lib.ImageExtras
 import com.gu.mediaservice.lib.config.UsageRightsConfig
 import com.gu.mediaservice.model._
+import lib.UsageQuota
 
 
-object CostCalculator {
+trait CostCalculator {
   import UsageRightsConfig.{freeSuppliers, suppliersCollectionExcl}
 
   val defaultCost = Pay
@@ -18,18 +18,20 @@ object CostCalculator {
   def isPay(usageRights: UsageRights): Boolean =
     getCost(usageRights) == Pay
 
+  def getOverQuota(usageRights: UsageRights) =
+    if (UsageQuota.isOverQuota(usageRights)) {
+      Some(Overquota)
+    } else {
+      None
+    }
+
   def getCost(usageRights: UsageRights): Cost = {
       val restricted  : Option[Cost] = usageRights.restrictions.map(r => Conditional)
       val categoryCost: Option[Cost] = usageRights.defaultCost
+      val overQuota: Option[Cost] = getOverQuota(usageRights)
       val supplierCost: Option[Cost] = usageRights match {
         case u: Agency => getCost(u.supplier, u.suppliersCollection)
         case _ => None
-      }
-
-      val overQuota: Option[Cost] = if (ImageExtras.isOverQuota(usageRights)) {
-        Some(Overquota)
-      } else {
-        None
       }
 
       restricted
